Add Cyberpunk Red, fecator some code
